refactor(markdown codec): export sync `encode` function
